requirements.txt
django
django-cors-headers
djangorestframework
requests
virustotal-python
dotenv
python-decouple
httpx
uvicorn
tldextract